remove pre-existing certs with the same nickname
When saving a certificate, once we know the new certificate's been
saved, delete any others with the same nickname.  If they have the same
nickname and subject, things mostly work, so we thought it'd be okay.
But sadly, various APIs and tools assume there's only ever one
certificate with a given nickname, so doing things like adjusting trust
settings and exporting certificates is still a bit too hit-or-miss.