set no-exit-on-disconnect when reconnecting
When we connect to the system bus, we set a flag on the connection so
that we don't exit if the connection is lost.  This allows us to attempt
to reconnect to the bus.  When reconnecting, we didn't set the same flag
on the new connection, so if we lost the connection a second time, the
D-Bus client library would automatically exit. (#948993)