Add lighter tests for DSA and EC
Add tests for EC and DSA.  For EC, we're only testing key generation,
not parameter generation.

Because both DSA and ECDSA signatures incorporate per-signature
randomness, we can't sign requests and certificates with both OpenSSL
and NSS and expect bit-for-bit identical results, so we don't bother -
the RSA versions of the tests already check that the content is right,
so we just check that the signatures verify.