Learn to multiplex possible CA tasks
Add enough machinery to be able to handle ongoing tasks on behalf of CA
entries as well as certificate monitoring/enrollment requests.  We don't
actually have any of those yet, but we will soon.

Also, clean up a bit of the memory handling logic when we're adding
either entries or CAs.