Force ASCII CN values in subject names to PrintableString
When encoding a common name value in a subject name, if it's ASCII,
force the encoding to PrintableString, which is what NSS uses.  Today's
OpenSSL update seems to always select UTF8String, which causes the same
input to result in different encodings of the same subject name
depending on which toolkit we use to do the encoding.