Handle "wait" status for CA data
When attempting to retrieve CA data, don't treat "wait" exit status
values as if we were indicating that the server was unreachable.
Instead, poll the helper for the information later (#1118468).
Also, if the server's unreachable, poll later at a fixed time even if
nothing interesting happens to the routing table.