Fix property Get and Set to use Variant wrapping
The implementation of property Get() and Set() didn't accept or return
values wrapped in a D-Bus Variant container, despite their introspection
signatures (which conform to the spec for the interface) claiming that
they did.  Use some of the recently-refactored message argument get/set
routines to fix that.  This expands on the original fix from David Kupka
so that it corrects more than just string values.