Also use a lock in system mode
Because we no longer always connect to the bus, we can't use whether or
not we are able to take the well-known service name as an indicator that
we're not stepping on another copy of the daemon.  Add a lock file to
the system daemon, as we already do for the session deamon.