Go back to retrying when cadata is unconfigured
We recently switched to disabling attempts at fetching root certificates
when the helper indicates that configuration for doing so is missing.
When IPA doesn't restart us at install-time, this means we don't start
noticing things until we get restarted for some other reason, which
isn't as friendly as we'd like to be.  Go back to the previous behavior
where we queued a refresh several hours in the future, but keep the new
path so that we can log the specific reason when debugging.