Also retrieve and cache an SCEP server's CA's cert
We need the CA's issuer's name to construct GetInitialCert messages, so
we need to cache that value as well.  Add logic to the helper to output
that second, if it exists, and tests that check if we're parsing data
from the helper right.  Also test that we cache CA capabilities
usefully.