Add a framework for decrypting enveloped-data
Add logic to the existing parse-everyting PKCS#7 routines to handle
decrypting enveloped-data.  The NSS version doesn't currently work
because the library only wants to use certificates in the database to
locate encryption keys, ignoring session certificates entirely, so it's
part of this patch only so that it can be picked up again if it's needed
for something else later.