Cache the last-transmitted SCEP nonce value
Instead of clearing SCEP values, cache the old nonce, and don't pass
them along to the helper if they're being re-used.  This lets the
SCEP-data-not-supplied path play out and cause new data to be generated,
while leaving the last-used data in the data file in case we need it for
troubleshooting.