Clean up the SCEP -R/-r options
Clean up how the SCEP helper handles the -R and -r options when it's
making GetCACert and GetCAChain requests.  Always hand back passed-in
values for the RA and CA certificates, if we have them, otherwise hand
back discovered values.  If we have passed-in values, add the discovered
values to the end of the possibly-in-the-chain list.