Document the dogtag helper's -N and -R flags
We already listed the -N flag in the help output, but it wasn't in the
man page, so add it: it reverses the default logic, whereby we would
normally prefer to attempt renewal if we have either an old certificate
(in $CERTMONGER_CERTIFICATE) or the serial number of an old certificate.
Also document the -R flag, which negates the -N flag.