Retrieve the list of profiles from Dogtag CAs
Add support for retrieving the list of available profiles from Dogtag
CAs.  For now, that's just so that we can track them.  Later on, we'll
teach getcert to warn about an unsupported value when the CA has some
vlaues, but it's been given a profile name that isn't among them.