Port "getcert" to popt
Switch to popt for parsing, but continue to use the old help messages
because the new text (while almost exactly the same) can't have been
translated yet.  Like the main program, add --autohelp/-H to evoke the
autohelp generator so that we can keep comparing it to the manual help
text.  Right now we don't use popt's automatic set-a-variable logic, but
we can tighten that up later.

We're attempting to use the same long option name for a given option
character across all of subcommands.

In places where we offer matching of an existing request using storage
locations, go ahead and accept token names for NSS databases, too (the
"refresh" and "status" commands).