Add some JSON support
Right now we're passing data around over pipes, and when that data is
complex, using some very ad-hoc formatting.  We don't have a readily
available JSON implementation in older releases, but we're going to need
to have one at some point for talking to CAs.  Being able to depend on
one being around so that we can use it for our own message passing would
be a serious improvement.