Start keeping track of key lifetimes and usage
Start tracking the dates on which we generated a key pair (if we did it)
and how many times we've sent requests to a CA using a given public key.
Also track the number of times we've saved certificates to disk as an
indication of how many times the CA has issued us a certificate for that
public key.