Accept CA roots as a JSON object
In addition to the hacky nicknames-and-certs lists, accept a JSON object
with elements named "roots", "other-roots", and "others", each of which
is an object with keys equal to suggested nicknames for use when storing
in an NSS database, and with their corresponding values being PEM-format
certificates.  Leave it undocumented for the moment until we're sure the
specific format works well enough.