Rework parsing of JSON enrollment results
Rework how we generate and parse the intermediate format that we're
using while parsing enrollment results to make it more extensible if we
ever end up needing to carry more than just the nickname and the PEM
certificate itself, and alter the test that passed this format out of
the helper to provide the new format.  The rest of the tests work as-is.