Document the helper interface from the helper PoV
The submit.txt document is written from the daemon's point of view, and
while that's great for maintainers who want a refresher on what the
daemon's supposed to be doing, it's not very helpful for people who want
to write enrollment helpers.

This document is meant to ultimately replace submit.txt as the
authoritative document -- if keeping this file and submit.txt in sync
becomes burdensome, we'll drop submit.txt.