Drop workarounds for DSA keygen with NSS
Don't artificially limit the size of DSA parameters that we try to
generate; instead, just let errors that we get when the requested size
is too big percolate up.
Don't bother trying to get certutil to generate larger DSA keys, since
it won't.
Don't expect to be able to generate 4096-bit keys, because softoken
won't.
Round up to the nearest 8-bit multiple when reporting key sizes, so that
our tests can accept keys with a few of the high bits not necessarily
set, since we're not screening them out any more.