Register our bus name after setting up handlers
Don't register the bus name until after we've set up the message filter
that passes requests to handlers, because being started by the bus means
that we need to be able to process requests that we receive before the
registration is acknowledged.  This should be the crucial part of making
ticket #38 work.