Tweak scep-submit's errors when it needs data
Tweak the error messages that scep-submit prints when it's not given
SCEP-specific data that it needs, to better describe what it expected.
The newer messages span multiple lines, which should also keep them from
being logged in submit-e.c, reducing confusion when it is part of the
normal flow of events (#1253009).