Document dogtag-ipa-renew-agent-submit's -t flag
Mention the -t flag in the man page for dogtag-ipa-renew-agent-submit,
to keep it in line with vanilla dogtag-submit.

At this point, the only things left that we don't note in the page for
dogtag-ipa-renew-agent-submit should be options related to how the tool
authenticates to the server, but since using IPA-agent-specific default
authentication settings is the point of having
dogtag-ipa-renew-agent-submit, that should be okay.