Fix a possible crash doing server location
When attempting server location to find the IPA server's RPC service, we
weren't checking that we had a DNS domain name before attempting to
either duplicate it or do IDN processing, which could cause a crash.
Add a NULL/empty check to catch that so that we can just return an
error.