ipa-submit: add options for using Basic auth
Add flags and logic to ipa-submit to allow a user name and password (or
password file) to be specified, and when they're specified, to switch
from using Kerberos to authenticate using the Negotiate mechanism, to
using passwords to authenticate using the Basic mechanism.  As this is
mainly intended for use as a troubleshooting option, the default is
still to use Kerberos creds obtained using a keytab.