Fix a flakiness in the 028-dbus test
When walking all of the exposed APIs from python, we were calling the
'resubmit' method on a certificate, and not waiting for the churn in
state that doing so would create to settle down before continuing.

This meant that the test script might have exited before the certmonger
process that was waiting on it finished saving the new certificate that
it obtained from resubmitting the rquest, so the process wouldn't
reliably log that it had obtained a new certificate.

Spotted by Jan Cholasta.