Handle X509 version being optional in tests
Newer versions of OpenSSL depart from older versions (and NSS) in that
they skip the version field when encoding a certificate where the field,
which is optional, has the default value of 0, indicating a v1
certificate.  To keep minicerts produced by both CSR generation
implementations the same, force the version to 3 (encoded value: 2) for
the tests.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>