Finish porting to OpenSSL 1.1
Check for new accessor functions, and wrap them in functions that
attempt to access the structure members directly when the accessor
functions aren't available.  In a couple of places, encode a structure
to DER, decode it with NSS, modify the NSS version of the structure,
re-encode it to DER, and re-decode it.  Learn to duplicate EVP_PKEYs by
encoding to DER and decoding.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>