Provide option to specify the EC curve, add EC default key size
The default RSA key size is 2048 which wasn't translated into an
equivalent EC key size which resulted in always using the 521-bit
curve.

Add an EC default key size of 256 as a correlation to the RSA
default (without adding a new curve).

The new option will allow one to specify the supported EC curves.

https://pagure.io/certmonger/issue/73

Signed-off-by: Rob Crittenden <rcritten@redhat.com>