MS cert template: add D-Bus property and storage
Add a D-Bus property and storage for a new template attribute, which
will hold data to be included in CSRs in the MS V2 certificate
template extension.

The server does not validate the extension data (validation
behaviour is implemented in a subsequent commit).

Part of: https://pagure.io/certmonger/issue/78