If stderr is not a tty log to syslog so the helpers can log
All the helpers were configured to use the log method cm_log_stderr
which when exececuted as a helper from the certmonger daemon would
log nowhere.

If stderr is detected as a tty (e.g. the helper is run directly on
the cli) then logging will go there. Otherwise it will log to
syslog (honoring the log level).