Remove NOMODDB flag flag from context init, look for full tokens
The NSS databases were almost universally initialized with the
NOMODDB flag. I'm not sure if something changed in NSS but the
PKCS#11 modules were not being initialized. Adding this back after
permission checks are done results in tokens working again.

When looking for certs and keys try the full token:nickname string
as well as just nickname when comparing values.

https://pagure.io/certmonger/issue/125