Adapt to the new behavior of disconnect in dbus-broker
The dbus server was replaced in Fedora-29 to use dbus-broker.

This server does not allow reconnects on a disconnect. certmonger
was crashing as a result.

The only way to directly pass a message between the dbus queue and
the main tevent handle is via a signal. So when a disconnect is
detected send a SIGTERM to getpid() in order to force a graceful
shutdown.

https://bugzilla.redhat.com/show_bug.cgi?id=1687698