Add long command-line options to man pages
The man pages almost universally only documented the short
options even though the long options were all defined in
the popt configuration.

Also do a bit of minor bit of reformatting and added a lint
option. I'm not going to require mandoc as a requirement as
the linting is pretty minor at the moment but it's better than
nothing.

https://bugzilla.redhat.com/show_bug.cgi?id=1782838