Drop hardcoded values for Apache NSS db for IPA < v4
These were used with IPAv3 when the IPA RA certificate
was stored in the Apache NSS database and references
by nickname. The RA certificate was moved to a set of
PEM files starting in IPA v4.

The hardcoded values were not particulary distribution
friendly.

This change isn't explicitly dropping support for IPAv3
but changes would be necessary there to pass in the
options for the NSS database directory and nickname.
A newer certmonger is not likely to be used with such an
old IPA release.

https://pagure.io/certmonger/issue/97

Signed-off-by: Rob Crittenden <rcritten@redhat.com>