Use the system env and not the session env for the local CA
The session env caused the local CA to try to use
/run/certmonger/.config/certmonger/certmonger.conf for its
configuration file which is both temporary and not exactly ideal.

Use the system environment as well so that
/etc/certmonger/certmonger.conf is the configuration file.

The problem was that users didn't know how to manage this file
and it didn't persist. The local CA is designed just for
developers but this wasn't discoverable at all and the
shipped certmonger.conf has a [local] section so confusion
abounded.

https://pagure.io/certmonger/issue/101